feat: display the mcp method + the mcp error code in the api logs detail page
https://gravitee.atlassian.net/browse/APIM-11430
