+++ gravitee-apim-console-webui/src/management/api/api-traffic-v4/runtime-logs-details/api-runtime-logs-details.component.html
@@ -1,13 +1,13 @@
 <!--
 
     Copyright (C) 2015 The Gravitee team (http://gravitee.io)
-    
+
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
     You may obtain a copy of the License at
-    
+
             http://www.apache.org/licenses/LICENSE-2.0
-    
+
     Unless required by applicable law or agreed to in writing, software
     distributed under the License is distributed on an "AS IS" BASIS,
     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,6 +20,6 @@
     <api-runtime-logs-messages></api-runtime-logs-messages>
   </ng-container>
   <ng-template #proxyBlock>
-    <api-runtime-logs-proxy></api-runtime-logs-proxy>
+    <api-runtime-logs-proxy [apiType]="apiType()"></api-runtime-logs-proxy>
   </ng-template>
 </ng-container>
