+++ gravitee-apim-console-webui/src/management/api/api-traffic-v4/runtime-logs-details/components/runtime-logs-proxy/api-runtime-logs-proxy.component.spec.ts
@@ -52,6 +52,7 @@ describe('ApiRuntimeLogsProxyComponent', () => {
     await TestBed.compileComponents();
     fixture = TestBed.createComponent(ApiRuntimeLogsProxyComponent);
     httpTestingController = TestBed.inject(HttpTestingController);
+    fixture.componentRef.setInput('apiType', 'PROXY');
     fixture.detectChanges();
 
     loader = TestbedHarnessEnvironment.loader(fixture);
