+++ gravitee-apim-rest-api/gravitee-apim-rest-api-model/src/main/java/io/gravitee/rest/api/model/v4/analytics/ApiMetricsDetail.java
@@ -18,6 +18,7 @@ package io.gravitee.rest.api.model.v4.analytics;
 import io.gravitee.common.http.HttpMethod;
 import io.gravitee.rest.api.model.v4.log.connection.ConnectionDiagnosticModel;
 import java.util.List;
+import java.util.Map;
 import lombok.Builder;
 import lombok.Data;
 
@@ -48,4 +49,5 @@ public class ApiMetricsDetail {
     String errorComponentName;
     String errorComponentType;
     List<ConnectionDiagnosticModel> warnings;
+    Map<String, String> additionalMetrics;
 }
