+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/analytics/use_case/FindApiMetricsDetailUseCase.java
@@ -87,6 +87,7 @@ public class FindApiMetricsDetailUseCase {
             .errorComponentName(apiMetricsDetail.getErrorComponentName())
             .errorComponentType(apiMetricsDetail.getErrorComponentType())
             .warnings(apiMetricsDetail.getWarnings())
+            .additionalMetrics(apiMetricsDetail.getAdditionalMetrics())
             .build();
 
         return new Output(result);
